pyproject.toml
[tool.poetry]
name = "ru-app"
version = "0.1.0"
description = "Service to connect database and mqtt."
authors = ["Victor Hugo Santos de Lima <vhsdl.eng19@uea.edu.br>"]

[tool.poetry.dependencies]
python = "^3.9"
motor = "^3.0.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
